Feat: 내 정보 조회 테스트
